% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccode_corpus_recode.R
\name{ccode_act}
\alias{ccode_act}
\title{function for aggregating corpus codes into topic categories}
\usage{
ccode_act(ccode)
}
\description{
function for aggregating corpus codes into topic categories
}

