% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccode_corpus_recode.R
\name{ccode_corpus_recode}
\alias{ccode_corpus_recode}
\title{function for aggregating corpus codes into topic categories}
\usage{
ccode_corpus_recode(ccode)
}
\description{
function for aggregating corpus codes into topic categories
}

