% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccode_corpus_recode.R
\name{ccode_top}
\alias{ccode_top}
\title{function for aggregating corpus codes into topic categories}
\usage{
ccode_top(ccode)
}
\description{
function for aggregating corpus codes into topic categories
}

