% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_link_dates.R
\name{check_link_dates}
\alias{check_link_dates}
\title{function for re-setting dates in linkage files}
\usage{
check_link_dates()
}
\description{
function for re-setting dates in linkage files
}

