% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{check_type_one}
\alias{check_type_one}
\title{checks for lines relevant=1 and ccode==999 at the same time}
\usage{
check_type_one(data_linkage, data_lines)
}
\description{
checks for lines relevant=1 and ccode==999 at the same time
}
\seealso{
\link{check_type_one} \link{check_type_two} \link{solve_type_one} \link{solve_type_two}
}

