% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tests.R
\name{check_type_two}
\alias{check_type_two}
\title{function for checking differences between linkage and corpus coding
relevant = 0 but ccode != 999}
\usage{
check_type_two(data_linkage, data_lines)
}
\description{
function for checking differences between linkage and corpus coding
relevant = 0 but ccode != 999
}
\seealso{
\link{check_type_one} \link{check_type_two} \link{solve_type_one} \link{solve_type_two}
}

