% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{classes}
\alias{classes}
\title{get classes}
\usage{
classes(x)
}
\arguments{
\item{x}{list or data frame}
}
\value{
returns list with class attributes of length equal to length(x)
}
\description{
get classes
}

