% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus_data_prepare.R
\name{corpus_data_prepare}
\alias{corpus_data_prepare}
\title{function for preparing corpus data to be merged with lines dat}
\usage{
corpus_data_prepare()
}
\description{
function for preparing corpus data to be merged with lines dat
}

