% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corpus_file_select.R
\name{corpus_file_select}
\alias{corpus_file_select}
\title{function for selecting a bunch of link files and computing other vars}
\usage{
corpus_file_select(corpus_file_full = NULL)
}
\description{
function for selecting a bunch of link files and computing other vars
}

