% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\docType{data}
\name{country_long}
\alias{country_long}
\title{country names matching the shorthands}
\format{\preformatted{ chr [1:16] "Austria" "Belgium" "Denmark" "Spain" "France" ...
}}
\usage{
country_long
}
\description{
country names matching the shorthands
}
\keyword{datasets}

