% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\docType{data}
\name{country_short}
\alias{country_short}
\title{country shorthands used throughout the project}
\format{An object of class \code{character} of length 16.}
\usage{
country_short
}
\description{
country shorthands used throughout the project
}
\keyword{datasets}

