% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ctest}
\alias{ctest}
\title{checks : check that the second text of version n
         is equal to the first text of version n+1
     ... which should be the case because these are
         supposed to be the same texts}
\usage{
ctest(link_texts, filelist_full)
}
\description{
checks : check that the second text of version n
         is equal to the first text of version n+1
     ... which should be the case because these are
         supposed to be the same texts
}

