% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{describe}
\alias{describe}
\title{function to describe variables in data.frame}
\usage{
describe(df, what = c("class", "nas", "min", "mean", "modus", "median",
  "max"), cols = TRUE)
}
\arguments{
\item{df}{data.frame to be described}

\item{what}{metric to put out}

\item{cols}{columns of data.frame for which to put out metrics}
}
\description{
function to describe variables in data.frame
}

