% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download_linelinkage_info}
\alias{download_linelinkage_info}
\title{function that downloads ready made info from the database}
\usage{
download_linelinkage_info(con)
}
\description{
function that downloads ready made info from the database
}

