% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download_textlines_raw}
\alias{download_textlines_raw}
\title{function that downloads ready made info from the database}
\usage{
download_textlines_raw(con, saveToFile = F)
}
\description{
function that downloads ready made info from the database
}

