% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_texts_raw.R
\name{download_texts_raw}
\alias{download_texts_raw}
\title{function that downloads ready made info from the database}
\usage{
download_texts_raw(con, saveToFile = F)
}
\description{
function that downloads ready made info from the database
}

