% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{email_error}
\alias{email_error}
\title{email error}
\usage{
email_error(..., s = "", body = "", anyways = F,
  to = "retep.meissner@gmail.com")
}
\arguments{
\item{s}{subject}

\item{body}{body of email}

\item{anyways}{if set to true function will mail also when in non-interactive mode while the default is to no mail}
}
\description{
email error
}

