% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{eval_cl_args}
\alias{eval_cl_args}
\title{function evaluating command line arguments to R that are in the form of a='value'}
\usage{
eval_cl_args()
}
\description{
function evaluating command line arguments to R that are in the form of a='value'
}

