% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_meta_from_fname}
\alias{get_meta_from_fname}
\title{function extracting dates and other information from a link file filename}
\usage{
get_meta_from_fname(fnames, flatten = F)
}
\description{
function extracting dates and other information from a link file filename
}

