% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_table_info.R
\name{get_table_info}
\alias{get_table_info}
\title{function that gatheres information on tables in a database}
\usage{
get_table_info(con, tables = "")
}
\description{
function that gatheres information on tables in a database
}

