% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{grepv}
\alias{grepv}
\title{function wrapper for grep(,value=TRUE)}
\usage{
grepv(pattern, x, ...)
}
\arguments{
\item{pattern}{regular expression to look for}

\item{x}{text to look through}

\item{...}{passed through to grep(pattern, x, value = TRUE, ...)}
}
\description{
function wrapper for grep(,value=TRUE)
}

