% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmltable.R
\name{htmltable}
\alias{htmltable}
\title{function for making html table}
\usage{
htmltable(x, file = "", standalone = TRUE, names = TRUE, head = "",
  foot = "", color = "", bgcolor = "", html = "", browse = file == "",
  width = rep("", dim(x)[2]))
}
\arguments{
\item{x}{data.frame or matrix to be transformed}

\item{file}{whether or not to write resulting HTML to file}

\item{standalone}{should HTML head and foot be added to make HTML standalone}

\item{names}{might be TRUE, FALSE or a character vector giving the names of the columns}

\item{head}{should user defined HTML head be used (everything before the table)}

\item{foot}{should user defined HTML foot be used (everything after the table)}
}
\description{
function for making html table
}

