% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_id}
\alias{is_id}
\title{function that determines if something might be an id (distinct and either integer or character)}
\usage{
is_id(x)
}
\description{
function that determines if something might be an id (distinct and either integer or character)
}

