% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data_check_textconsistency.R
\name{link_data_check_textconsistency}
\alias{link_data_check_textconsistency}
\title{function that checks whether text b of linkage 1 and text a of linkage 2 are identical}
\usage{
link_data_check_textconsistency(link_texts, filelist_full)
}
\description{
function that checks whether text b of linkage 1 and text a of linkage 2 are identical
}

