% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/link_data_check_textlength.R
\name{link_data_check_textlength}
\alias{link_data_check_textlength}
\title{function that checks the lengths of texts}
\usage{
link_data_check_textlength(textlist_flat)
}
\description{
function that checks the lengths of texts
}

