% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data_clean_text.R
\name{link_data_clean_text}
\alias{link_data_clean_text}
\title{function that cleans up text from link data}
\usage{
link_data_clean_text(text)
}
\description{
function that cleans up text from link data
}

