% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/link_data_get_date.R
\name{link_data_get_date}
\alias{link_data_get_date}
\title{function for extracting dates from link data texts}
\usage{
link_data_get_date(RESULTS, flatten = F)
}
\description{
function for extracting dates from link data texts
}

