% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data_get_text.R
\name{link_data_get_text}
\alias{link_data_get_text}
\title{function extracting the texts from link data}
\usage{
link_data_get_text(RESULTS, both = T)
}
\description{
function extracting the texts from link data
}

