% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/link_files_get_date.R
\name{link_files_get_date}
\alias{link_files_get_date}
\title{function for extracting dates from link data files}
\usage{
link_files_get_date(filelist_full, flatten = F)
}
\description{
function for extracting dates from link data files
}

