% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_files_get_linkage.R
\name{link_files_get_linkage}
\alias{link_files_get_linkage}
\title{function for extracting dates from link data files}
\usage{
link_files_get_linkage()
}
\description{
function for extracting dates from link data files
}

