% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/link_files_get_text.R
\name{link_files_get_text}
\alias{link_files_get_text}
\title{function extracting the texts from link data files}
\usage{
link_files_get_text(linkage_env)
}
\description{
function extracting the texts from link data files
}

