% Generated by roxygen2 (4.0.2): do not edit by hand
\name{link_files_get_text_only}
\alias{link_files_get_text_only}
\title{function extracting the texts from link data files}
\usage{
link_files_get_text_only(filelist_full, flatten = F)
}
\description{
function extracting the texts from link data files
}

