% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_files_load.R
\name{link_files_load}
\alias{link_files_load}
\title{function for loading a set off link data files}
\usage{
link_files_load(filelist_full)
}
\description{
function for loading a set off link data files
}

