% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/link_files_select.R
\name{link_files_select}
\alias{link_files_select}
\title{function for selecting a bunch of link files and computing other vars}
\usage{
link_files_select(filelist_full = NULL)
}
\description{
function for selecting a bunch of link files and computing other vars
}

