% Generated by roxygen2 (4.0.2): do not edit by hand
\name{linkage_sim}
\alias{linkage_sim}
\title{function for calculating similarity again}
\usage{
linkage_sim(res, simtype = c("sim", "diff", "sim_wd", "diff_wd", "wds", "wd1",
  "wd2"))
}
\arguments{
\item{simtype}{is one of sim, diff, sim_wd, diff_wd, wds, wd1, wd2}
}
\description{
function for calculating similarity again
}

