% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linkage_to_html.R
\name{linkage_to_html}
\alias{linkage_to_html}
\title{present linkage information in HTML file}
\usage{
linkage_to_html(linkagedata, out = "")
}
\description{
present linkage information in HTML file
}

