% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_type_colors.R
\name{linkage_type_colors}
\alias{linkage_type_colors}
\title{function to determine colors based on linkage type}
\usage{
linkage_type_colors(type)
}
\arguments{
\item{type}{character vector with one of the following: 
c("no-change", "insertion", "change", "deletion")}
}
\description{
function to determine colors based on linkage type
}

