% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ll_view_write_texts.R
\name{ll_view_write_texts}
\alias{ll_view_write_texts}
\title{function producing text / html versions of standing order texts}
\usage{
ll_view_write_texts(df)
}
\description{
function producing text / html versions of standing order texts
}

