% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have_match.R
\name{matched}
\alias{matched}
\title{function for looking which items of a are contained in b}
\usage{
matched(a, b)
}
\arguments{
\item{a}{items to be checked}

\item{b}{items to check against}
}
\description{
function for looking which items of a are contained in b
}

