% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{maxchar}
\alias{maxchar}
\title{function for restricting size of string}
\usage{
maxchar(text, maxchar = 5, indicator = " ...")
}
\arguments{
\item{text}{vector of text}
}
\description{
function for restricting size of string
}

