% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger_for_time_spans.R
\name{mfts_match_dates_to_id}
\alias{mfts_match_dates_to_id}
\title{matching ids to dates}
\usage{
mfts_match_dates_to_id(full_day_sequence, df)
}
\arguments{
\item{full_day_sequence}{a sequence of days span ids should be matched to}

\item{df}{a data frame containing span inforamtion with}
}
\description{
function matching a sequence of dates and a data frame of id, start, end
}

