% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_var_overview}
\alias{plot_var_overview}
\title{function plotting the range of a variable}
\usage{
plot_var_overview(x)
}
\arguments{
\item{x}{varaible to be plotted}
}
\description{
function plotting the range of a variable
}

