% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report_time.R
\name{report_time}
\alias{report_time}
\title{function for reporting the time it took to execute statement}
\usage{
report_time(expr)
}
\description{
function for reporting the time it took to execute statement
}

