% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reset_link_dates.R
\name{reset_link_dates}
\alias{reset_link_dates}
\title{function for re-setting dates in linkage files}
\usage{
reset_link_dates()
}
\description{
function for re-setting dates in linkage files
}

