% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqalong.R
\name{seqalong}
\alias{seqalong}
\title{convenience function for getting seqence for looping}
\usage{
seqalong(x)
}
\arguments{
\item{x}{the object to form a sequence along}
}
\value{
a vector of length (or first dimension) of x
}
\description{
convenience function for getting seqence for looping
}

