% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_enc_utf8.R
\name{set_enc_utf8}
\alias{set_enc_utf8}
\title{function setting Encoding value of string vectors in data frame to utf-8}
\usage{
set_enc_utf8(df)
}
\arguments{
\item{df}{data.frame to be cured}
}
\description{
function setting Encoding value of string vectors in data frame to utf-8
}

