% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{solve_type_one}
\alias{solve_type_one}
\title{solve where lines relevant=1 and ccode==999 at the same time
solution: 
- set relevant to 0 because ccoding is wiser
- delete linkage information for this line because it is not relevant!}
\usage{
solve_type_one(check_type_one_results)
}
\description{
solve where lines relevant=1 and ccode==999 at the same time
solution: 
- set relevant to 0 because ccoding is wiser
- delete linkage information for this line because it is not relevant!
}
\seealso{
\link{check_type_one} \link{check_type_two} \link{solve_type_one} \link{solve_type_two}
}

