% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tests.R
\name{solve_type_two}
\alias{solve_type_two}
\title{function for solving problems where ccode != 999 but rel = 0
solution:
- relevant <- 1 (ccode is wiser)
- linkage1 <- deletion (add forward conservative forward linking)
- linkage2 <- insertion (add conservative backward linking)
- minmaj   <- 0 (set minmaj conservatively to 0)}
\usage{
solve_type_two(check_type_two_results)
}
\description{
function for solving problems where ccode != 999 but rel = 0
solution:
- relevant <- 1 (ccode is wiser)
- linkage1 <- deletion (add forward conservative forward linking)
- linkage2 <- insertion (add conservative backward linking)
- minmaj   <- 0 (set minmaj conservatively to 0)
}
\seealso{
\link{check_type_one} \link{check_type_two} \link{solve_type_one} \link{solve_type_two}
}

