% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sort_alignment_df.r
\name{sort_align_dev}
\alias{sort_align_dev}
\title{function sorting alignment data frames according to line numbers}
\usage{
sort_align_dev(df, first, second, dnlb = NULL)
}
\arguments{
\item{df}{the alignment data frame to be sorted}

\item{first}{name of first column to be sorted}

\item{dnlb}{a special parameter for sorting NULL per default.
Should be either NULL or a vector of size dim(df)[1].
If set those lines/rows marked will be tried to put behind second
predecessors}

\item{first}{name of second column to be sorted}
}
\description{
function sorting alignment data frames according to line numbers
}

