% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sort_alignment_df.r
\name{sort_align_df}
\alias{sort_align_df}
\title{function sorting alignment data frames according to line numbers}
\usage{
sort_align_df(df, first, second)
}
\arguments{
\item{df}{the alignment data frame to be sorted}

\item{first}{name of first column to be sorted}

\item{first}{name of second column to be sorted}
}
\description{
function sorting alignment data frames according to line numbers
}

