% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger_for_time_spans.R
\name{split_timespan_after}
\alias{split_timespan_after}
\title{function splitting sequence after certain date}
\usage{
split_timespan_after(start, end, splitdate, splitdateold = NA)
}
\arguments{
\item{start}{vector of start values}

\item{end}{vector of end values}

\item{splidate}{data after which time span shall be split}
}
\description{
function splitting sequence after certain date
}

