% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text_dist.R
\name{text_dist}
\alias{text_dist}
\title{function for calculating the distance between two text snippets}
\usage{
text_dist(text1, text2, clean = F)
}
\description{
function for calculating the distance between two text snippets
}

