% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccode_corpus_recode.R
\docType{data}
\name{ccode_color_scheme}
\alias{ccode_color_scheme}
\title{coler scheme}
\format{\preformatted{ chr [1:7] "#ec7014" "#fcbba1" "#D94801" "#41ab5d" "#e7298a" ...
}}
\usage{
ccode_color_scheme
}
\description{
coler scheme
}
\keyword{datasets}

