% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ccode_color_scheme}
\alias{ccode_color_scheme}
\title{function for recoding corpus code aggregation scheme (late 2014)}
\format{\preformatted{ chr [1:7] "#ec7014" "#fcbba1" "#D94801" "#41ab5d" "#e7298a" ...
}}
\usage{
ccode_color_scheme
}
\description{
function for recoding corpus code aggregation scheme (late 2014)
}
\keyword{datasets}

