% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccode_corpus_recode.R
\docType{data}
\name{ccode_color_scheme2}
\alias{ccode_color_scheme2}
\title{coler scheme}
\format{An object of class \code{character} of length 7.}
\usage{
ccode_color_scheme2
}
\description{
coler scheme
}
\keyword{datasets}

