% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccode_corpus_recode.R
\docType{data}
\name{ccode_color_scheme2}
\alias{ccode_color_scheme2}
\title{coler scheme}
\format{\preformatted{ chr [1:7] "#F3FFE8" "#FFFAEF" "#F6E9F2" "#F4F4FF" "#FEFBE8" ...
}}
\usage{
ccode_color_scheme2
}
\description{
coler scheme
}
\keyword{datasets}

