% Generated by roxygen2 (4.0.2): do not edit by hand
\name{corpus_file_select}
\alias{corpus_file_select}
\title{function for selecting a bunch of link files and computing other vars}
\usage{
corpus_file_select()
}
\description{
function for selecting a bunch of link files and computing other vars
}

