% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\docType{data}
\name{country_long}
\alias{country_long}
\title{country names matching the shorthands}
\format{An object of class \code{character} of length 16.}
\usage{
country_long
}
\description{
country names matching the shorthands
}
\keyword{datasets}

