% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\docType{data}
\name{country_short}
\alias{country_short}
\title{country shorthands used throughout the project}
\format{\preformatted{ chr [1:16] "AUT" "BEL" "DEN" "ESP" "FRA" "GER" "IRE" "ITA" ...
}}
\usage{
country_short
}
\description{
country shorthands used throughout the project
}
\keyword{datasets}

