% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{desc_df}
\alias{desc_df}
\title{function to describe variables in data.frame}
\usage{
desc_df(df, what = c("class", "nas", "min", "mean", "modus", "median", "max"),
  cols = TRUE, digits = 2)
}
\arguments{
\item{df}{data.frame to be described}

\item{what}{metric to put out}

\item{cols}{columns of data.frame for which to put out metrics}

\item{digits}{shall numbers be rounded to a certain amount of diogits?}
}
\description{
function to describe variables in data.frame
}

