% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_linelinkage_raw.R
\name{download_linelinkage_raw}
\alias{download_linelinkage_raw}
\title{function that downloads ready made info from the database}
\usage{
download_linelinkage_raw(con, saveToFile = F)
}
\description{
function that downloads ready made info from the database
}

