% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_meta_from_fname.R
\name{get_meta_from_fname}
\alias{get_meta_from_fname}
\title{function extracting dates and other information from a link file filename}
\usage{
get_meta_from_fname(fnames, flatten = F, uniq = T)
}
\description{
function extracting dates and other information from a link file filename
}

