% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data_get_linkage.R
\name{link_data_get_linkage}
\alias{link_data_get_linkage}
\title{function extracting linkage information}
\usage{
link_data_get_linkage(res, meta)
}
\description{
function extracting linkage information
}

