% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/link_files_get_text_only.R
\name{link_files_get_text_only}
\alias{link_files_get_text_only}
\title{function extracting the texts from link data files}
\usage{
link_files_get_text_only(filelist_full, flatten = F)
}
\description{
function extracting the texts from link data files
}

