% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modus.R
\name{modus}
\alias{modus}
\title{function for getting mode source: http://stackoverflow.com/a/8189441/1144966}
\usage{
modus(x)
}
\arguments{
\item{x}{vector for which the mode should be returned}
}
\description{
function for getting mode source: http://stackoverflow.com/a/8189441/1144966
}

