% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{sqlEscape}
\alias{sqlEscape}
\title{Some funtions making life easier for handling Databases from within R}
\usage{
sqlEscape(data)
}
\description{
Some funtions making life easier for handling Databases from within R
}

